fix(transform): concurrent now works
The old concurrency model was copied from promise stream and had some issues with events finishing at different times. We now use a set to figure out what's still in flight and let promises remove themselves, no timing required.
