+++ lib/read.js
@@ -24,7 +24,7 @@ module.exports = class PromiseReadStream extends ReadableStream {
     this._keepReading = false
     this._streamDeferred = defer()
     this.wrapRead()
-    this.once('end', this._finishUp)
+    this.once('end', () => this._streamDeferred.resolve())
   }
 
   wrapRead () {
@@ -60,10 +60,6 @@ module.exports = class PromiseReadStream extends ReadableStream {
 
   emitError (e) {
     this.emit('error', e)
-  }
-
-  _finishUp () {
-    this._streamDeferred.resolve()
   }
 
   promise () {
