+++ lib/transform.js
@@ -2,12 +2,6 @@ const Promise = require('bluebird')
 const Transform = require('stream').Transform
 const { defer, maybeResume } = require('./utils')
 
-function nextTick () {
-  return new Promise(function (resolve, reject) {
-    process.nextTick(resolve)
-  })
-}
-
 module.exports = class PromiseTransformStream extends Transform {
   constructor (opts, fn) {
     if (typeof opts === 'function') {
@@ -34,36 +28,30 @@ module.exports = class PromiseTransformStream extends Transform {
     this._streamEnd = defer()
     this._handlingErrors = false
     this._concurrent = opts.concurrent
-    this._queue = []
+    this._queue = new Set()
   }
 
   wrapTransform () {
     const _fn = this._transform
-    this._transform = (data, encoding, done) => {
-      var queue = this._queue
-      var processed = Promise.resolve()
+    this._transform = function (data, encoding, done) {
+      const processed = Promise.resolve()
         .then(() => _fn.call(this, data, encoding))
         .then(data => {
           if (data !== undefined) {
-            this.push(data)
+            return this.push(data)
           }
         })
-      processed.catch(done)
-      queue.push(processed)
-      if (queue.length >= this._concurrent) {
-        var next = queue.shift()
-        // The delay is a workaround for the bad design of
-        // node streams which forbid you to call done twice
-        // at the same tick on the event loop, even if you
-        // had events happening at the exact same tick
-        if (next.isResolved()) {
-          nextTick().then(done, done)
-        } else {
-          next.then(done, done)
-        }
-      } else {
-        done()
+        .then(() => {
+          this._queue.delete(processed)
+        }, e => {
+          this.emitError(e)
+        })
+      this._queue.add(processed)
+
+      if (this._queue.size < this._concurrent) {
+        return done()
       }
+      Promise.race(this._queue).then(() => done(), e => this.emitError(e))
     }
   }
 
@@ -71,7 +59,7 @@ module.exports = class PromiseTransformStream extends Transform {
     const flushCb = this._flush
     this._flush = (done) => {
       Promise.all(this._queue)
-        .then(() => this._finishUp())
+        .then(() => this._streamEnd.resolve())
         .then(() => {
           if (flushCb) {
             return flushCb.call(this)
@@ -81,14 +69,12 @@ module.exports = class PromiseTransformStream extends Transform {
     }
   }
 
-  _finishUp () {
-    this._streamEnd.resolve()
-  }
-
   push (data) {
     return Promise.resolve(data)
-      .bind(this)
-      .then(super.push, this.emitError)
+      .then(
+        data => super.push(data),
+        err => this.emitError(err)
+      )
   }
 
   emitError (error) {
