+++ lib/utils.js
@@ -4,7 +4,7 @@ function defer () {
     resolveCb = resolve
     rejectCb = reject
   })
-  return { resolve: resolveCb, reject: rejectCb, promise: promise }
+  return { resolve: resolveCb, reject: rejectCb, promise }
 }
 
 function maybeResume (stream) {
